[symex] Move thread interleaving into goto-symex/scheduler/
Sixth of seven commits partitioning src/goto-symex/
(docs/roadmap/goto-symex-partition-plan.md §2). execution_state and
reachability_tree are the only files that know there is more than one thread.

Pure move apart from one prose citation: LTL.md's execution_state line number
was stale before this PR and is re-derived here.
