+++ src/goto-symex/builtin_functions/object_size.cpp
@@ -1,6 +1,6 @@
 #include <cassert>
 #include <goto-symex/goto_symex.h>
-#include <goto-symex/reachability_tree.h>
+#include <goto-symex/scheduler/reachability_tree.h>
 #include <util/arith/arith_tools.h>
 #include <util/lang/c_types.h>
 #include <util/expr/expr_util.h>
