+++ src/goto-symex/state/goto_symex_state.cpp
@@ -1,8 +1,8 @@
 #include <cassert>
-#include <goto-symex/execution_state.h>
+#include <goto-symex/scheduler/execution_state.h>
 #include <goto-symex/goto_symex.h>
 #include <goto-symex/state/goto_symex_state.h>
-#include <goto-symex/reachability_tree.h>
+#include <goto-symex/scheduler/reachability_tree.h>
 #include <map>
 #include <optional>
 #include <sstream>
