+++ src/goto-symex/symex_assign.cpp
@@ -1,7 +1,7 @@
 #include <cassert>
 #include <climits>
 #include <goto-symex/dynamic_allocation.h>
-#include <goto-symex/execution_state.h>
+#include <goto-symex/scheduler/execution_state.h>
 #include <goto-symex/goto_symex.h>
 #include <util/lang/c_types.h>
 #include <util/symtab/cprover_prefix.h>
