+++ src/goto-symex/symex_function.cpp
@@ -1,7 +1,7 @@
 #include <algorithm>
 #include <cassert>
 #include <unordered_set>
-#include <goto-symex/execution_state.h>
+#include <goto-symex/scheduler/execution_state.h>
 #include <goto-symex/goto_symex.h>
 #include <langapi/language_util.h>
 #include <util/arith/arith_tools.h>
