+++ src/goto-symex/symex_main.cpp
@@ -1,9 +1,9 @@
 #include <cassert>
 #include <optional>
-#include <goto-symex/execution_state.h>
+#include <goto-symex/scheduler/execution_state.h>
 #include <goto-symex/goto_symex.h>
 #include <goto-symex/state/goto_symex_state.h>
-#include <goto-symex/reachability_tree.h>
+#include <goto-symex/scheduler/reachability_tree.h>
 #include <goto-symex/equation/symex_target_equation.h>
 
 #include <langapi/language_util.h>
