+++ unit/goto-symex/context_stack.test.cpp
@@ -37,7 +37,7 @@
 #include <string>
 #include <utility>
 
-#include <goto-symex/reachability_tree.h>
+#include <goto-symex/scheduler/reachability_tree.h>
 #include <goto-symex/equation/symex_target_equation.h>
 #include <solvers/smt/smt_conv.h>
 #include <solvers/solve.h>
