+++ unit/goto-symex/mpor.test.cpp
@@ -31,7 +31,7 @@
 #include <string>
 #include <vector>
 
-#include <goto-symex/reachability_tree.h>
+#include <goto-symex/scheduler/reachability_tree.h>
 #include <goto-symex/equation/symex_target_equation.h>
 #include <util/symtab/namespace.h>
