+++ unit/goto-symex/renaming.test.cpp
@@ -31,7 +31,7 @@
 
 #include <string>
 
-#include <goto-symex/reachability_tree.h>
+#include <goto-symex/scheduler/reachability_tree.h>
 #include <goto-symex/state/renaming.h>
 #include <irep2/irep2_expr.h>
 #include <util/lang/c_types.h>
