+++ unit/goto-symex/ssa_wellformed.test.cpp
@@ -19,7 +19,7 @@
 #include <algorithm>
 #include <string>
 
-#include <goto-symex/reachability_tree.h>
+#include <goto-symex/scheduler/reachability_tree.h>
 #include <goto-symex/equation/symex_target_equation.h>
 #include <util/symtab/namespace.h>
