+++ unit/goto-symex/symex_run.h
@@ -29,7 +29,7 @@
 
 #include <goto-programs/goto_k_induction.h>
 #include <goto-programs/remove_no_op.h>
-#include <goto-symex/reachability_tree.h>
+#include <goto-symex/scheduler/reachability_tree.h>
 #include <goto-symex/equation/symex_target_equation.h>
 #include <util/symtab/namespace.h>
