+++ unit/goto-symex/unwind.test.cpp
@@ -21,7 +21,7 @@
 #include <string>
 
 #include <goto-programs/goto_functions.h>
-#include <goto-symex/reachability_tree.h>
+#include <goto-symex/scheduler/reachability_tree.h>
 #include <irep2/irep2_expr.h>
 #include <util/symtab/namespace.h>
